+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 ### Added
 - Added the “Captions/Subtitles” file kind. ([#7304](https://github.com/craftcms/cms/issues/7304))
 - Added the `handleCasing` config setting, which determines the default casing that should be used when autogenerating component handles. ([#4276](https://github.com/craftcms/cms/issues/4276))
+- Added `craft\base\ApplicationTrait::getUnbufferedDb()`.
 - Added `craft\helpers\Template::paginateQuery()`.
 
 ### Changed
 - The `allowedFileExtensions` config setting now includes several file extensions used by caption and subtitle file formats by default. ([#7304](https://github.com/craftcms/cms/issues/7304))
+- `craft\db\Query::batch()` and `each()` now use an [unbuffered database connection](https://www.php.net/manual/en/mysqlinfo.concepts.buffering.php) by default if using MySQL. ([#7338](https://github.com/craftcms/cms/issues/7338))
 - Craft no longer reports user deprecation errors logged with `E_USER_DEPRECATED`.
 
 ### Deprecated